refactor(realtime): clear branch-CI style gate and pin replay proof to legacy
The changed-style gate rejects worsened file-length and layout findings, so
the damping additions that had grown pre-existing large files move into
feature modules, byte-reverting their hosts to main size or below:

- state-sync/state-sync-payload.ts owns state-sync payload decoding
  (parse + guards + readGroupSnapshotStateSyncPayload) out of
  state-sync-routing.ts; state-sync/validate-state-sync.ts owns the
  computed-state-sync validators out of state-sync-publisher.ts.
- topology/rallar-rtc-topology-metrics.ts owns the topology metrics
  contract and zero factory; the dead per-service recompute-debounce
  option/reader is removed (the env var now resolves through
  readApiTopologyRecomputeDebounceMs and the coalesced-work module's
  DEFAULT_TOPOLOGY_RECOMPUTE_DEBOUNCE_MS).
- topology/replay/finish-rtc-topology-work.ts owns the reservation-fenced
  completion out of the work handler; the coalesced group-revision work
  module moves to topology/replay/ beside its consumer and now owns
  isChangeGatedGroupRevisionWork.
- postgres/resource-inbox/resource-inbox-finished-replacement.ts owns the
  terminal-revival CAS as a standalone function called with the coalescer's
  write transaction; ResourceInboxRepository is byte-identical to main.
- al-contracts/read-al-principal-broadcast-target.ts owns the principal
  broadcast-target reader out of the al-contract protocol module.
- api-v1 middleware wiring flows through readApiGroupFormationTopologyIntent;
  structural-lineage manifests map the extractions for the changed checker.

Both deterministic topology replay proof steps (standalone gate and release
gate) now run with RALLAR_GROUP_FORMATION_DAMPING=legacy: the proof pins the
legacy per-command publication contract (one durable append per mutation,
per-command publication ids) that the damped default intentionally retires;
the durable replay machinery it proves is unchanged. Adapting the proof to
the damped contract is tracked as a follow-up issue.

Recipe-structure tests pin the reshaped convergence recipe (state-plane
snapshot waits and exact-revision asserts), and the handler-shape test
follows the reservation completion into its module without weakening the
fenced-transaction proof.
