+++ packages/shared-server/rallar-system/services/rallar-rtc-topology-service.ts
@@ -8,6 +8,10 @@ import {
     type RallarRtcTopologyKind,
 } from '@shared/api/overlay-topology.ts';
 import { rtcTopologySemanticEqual } from '../rtc-topology-semantic-equality.ts';
+import {
+    emptyTopologyMetrics,
+    type RallarRtcTopologyMetrics,
+} from '../topology/rallar-rtc-topology-metrics.ts';
 import { normalizeRttReportingDegreeLimit } from '@shared/rtc/rtt-reporting-policy.ts';
 import {
     compareRtcTopologyIdentifiers,
@@ -43,43 +47,7 @@ export type RallarRtcTopologyServiceOptions = Readonly<{
     meshMinSize?: number;
     meshParamK?: number;
     rttRebuildDebounceMs?: number;
-    topologyRecomputeDebounceMs?: number;
     now?: () => number;
-}>;
-
-export type RallarRtcTopologyMetrics = Readonly<{
-    topologyUpdateCount: number;
-    topologyChangedCount: number;
-    topologyUnchangedCount: number;
-    updatesWithRttMeasurementCount: number;
-    updatesWithoutRttMeasurementCount: number;
-    starPlanCount: number;
-    starPlanDurationMs: number;
-    noRttTreePlanCount: number;
-    noRttTreePlanDurationMs: number;
-    noRttMeshPlanCount: number;
-    noRttMeshPlanDurationMs: number;
-    weightedPlanCount: number;
-    weightedPlanDurationMs: number;
-    weightedRoomGraphBuildCount: number;
-    weightedRoomGraphBuildDurationMs: number;
-    weightedRoomGraphSparseFallbackCount: number;
-    rttQueueRequestCount: number;
-    rttQueueNewCount: number;
-    rttQueueCoalescedCount: number;
-    rttQueueImmediateCount: number;
-    rttFlushAttemptCount: number;
-    rttFlushSkippedCount: number;
-    rttFlushExecutedCount: number;
-    topologyPublishAttemptCount: number;
-    topologyPublishedCount: number;
-    topologyPublishSkippedUnchangedCount: number;
-    topologyRebuildSkippedFingerprintCount: number;
-    topologyRemovalRequestCount: number;
-    topologyRemovedCount: number;
-    topologyRemoveMissCount: number;
-    topologySnapshotCount: number;
-    pendingRttUpdateCount: number;
 }>;
 
 export type RallarRtcTopologyUpdateResult = Readonly<{
@@ -160,47 +128,6 @@ const DEFAULT_TREE_MIN_SIZE = 5;
 const DEFAULT_MESH_MIN_SIZE = 16;
 const DEFAULT_MESH_PARAM_K = 2;
 const DEFAULT_RTT_REBUILD_DEBOUNCE_MS = 250;
-export const DEFAULT_TOPOLOGY_RECOMPUTE_DEBOUNCE_MS = 500;
-
-type MutableRallarRtcTopologyMetrics = {
-    -readonly [K in keyof Omit<
-        RallarRtcTopologyMetrics,
-        'pendingRttUpdateCount' | 'topologySnapshotCount'
-    >]: number;
-};
-
-const emptyTopologyMetrics = (): MutableRallarRtcTopologyMetrics => ({
-    topologyUpdateCount: 0,
-    topologyChangedCount: 0,
-    topologyUnchangedCount: 0,
-    updatesWithRttMeasurementCount: 0,
-    updatesWithoutRttMeasurementCount: 0,
-    starPlanCount: 0,
-    starPlanDurationMs: 0,
-    noRttTreePlanCount: 0,
-    noRttTreePlanDurationMs: 0,
-    noRttMeshPlanCount: 0,
-    noRttMeshPlanDurationMs: 0,
-    weightedPlanCount: 0,
-    weightedPlanDurationMs: 0,
-    weightedRoomGraphBuildCount: 0,
-    weightedRoomGraphBuildDurationMs: 0,
-    weightedRoomGraphSparseFallbackCount: 0,
-    rttQueueRequestCount: 0,
-    rttQueueNewCount: 0,
-    rttQueueCoalescedCount: 0,
-    rttQueueImmediateCount: 0,
-    rttFlushAttemptCount: 0,
-    rttFlushSkippedCount: 0,
-    rttFlushExecutedCount: 0,
-    topologyPublishAttemptCount: 0,
-    topologyPublishedCount: 0,
-    topologyPublishSkippedUnchangedCount: 0,
-    topologyRebuildSkippedFingerprintCount: 0,
-    topologyRemovalRequestCount: 0,
-    topologyRemovedCount: 0,
-    topologyRemoveMissCount: 0,
-});
 
 export class RallarRtcTopologyService {
     private readonly snapshotsByOverlayId = new Map<
@@ -475,14 +402,6 @@ export class RallarRtcTopologyService {
 
     readRttRebuildDebounceMs(): number {
         return this.rttRebuildDebounceMs();
-    }
-
-    readTopologyRecomputeDebounceMs(): number {
-        return Math.max(
-            0,
-            this.options.topologyRecomputeDebounceMs ??
-                DEFAULT_TOPOLOGY_RECOMPUTE_DEBOUNCE_MS,
-        );
     }
 
     readNowEpochMs(): number {
