+++ packages/shared/al-contracts/al-contract.ts
@@ -44,12 +44,7 @@ export type ALTargets =
     mode: 'broadcast';
     scope: 'room' | 'world' | 'all' | 'principal';
     groupRef?: GroupRef;
-    /**
-     * Server-resolved audience identity for scope 'principal': the principal's
-     * own live sessions plus live sessions of groups the principal is an
-     * active member of. Never resolves to every open connection.
-     */
-    principalRef?: ClientPrincipalRef;
+    principalRef?: ClientPrincipalRef; // scope 'principal': own + co-group live sessions only
     exceptPeerIds?: readonly string[];
     minSnapshotVersion?: number;
     /** Immutable logical audience captured by authoritative server work. */
@@ -338,9 +333,7 @@ export function toALGroupTargetKey(group: string | GroupRef): string {
     ]);
 }
 
-export function readALMulticastTargetGroupRef(
-    message: ALMessage,
-): GroupRef | undefined {
+export function readALMulticastTargetGroupRef(message: ALMessage): GroupRef | undefined {
     const targets = message.targets;
     if (targets?.mode !== 'multicast') {
         return undefined;
@@ -349,31 +342,7 @@ export function readALMulticastTargetGroupRef(
     return toALGroupRef(targets.groupRef);
 }
 
-export function readALPrincipalBroadcastTarget(
-    message: ALMessage,
-): ClientPrincipalRef | undefined {
-    const targets = message.targets;
-    if (
-        targets?.mode !== 'broadcast' ||
-        targets.scope !== 'principal' ||
-        targets.principalRef === undefined ||
-        typeof targets.principalRef.applicationId !== 'string' ||
-        typeof targets.principalRef.workspaceId !== 'string' ||
-        typeof targets.principalRef.principalId !== 'string'
-    ) {
-        return undefined;
-    }
-
-    return {
-        applicationId: targets.principalRef.applicationId,
-        workspaceId: targets.principalRef.workspaceId,
-        principalId: targets.principalRef.principalId,
-    };
-}
-
-export function readALTargetGroupRef(
-    message: ALMessage,
-): GroupRef | undefined {
+export function readALTargetGroupRef(message: ALMessage): GroupRef | undefined {
     const targets = message.targets;
     if (targets?.mode === 'multicast') {
         return toALGroupRef(targets.groupRef);
