+++ plans/repo-style-lineages/group-formation-phase2-server-damping.json
@@ -1,0 +1,45 @@
+{
+  "version": 1,
+  "lineages": [
+    {
+      "mergeBase": "1e5f5e55e6ff94c016bfe2cc11af92952a30e32f",
+      "source": {
+        "path": "packages/shared-server/rallar-system/state-sync-routing.ts",
+        "blob": "b103194a15bbf2af3999106e84b53d49539942c6"
+      },
+      "targets": [
+        "packages/shared-server/rallar-system/state-sync/state-sync-payload.ts"
+      ]
+    },
+    {
+      "mergeBase": "1e5f5e55e6ff94c016bfe2cc11af92952a30e32f",
+      "source": {
+        "path": "packages/shared-server/rallar-system/state-sync-publisher.ts",
+        "blob": "12beeaf06106a06fcf2a9c67bb49d9093cea60c4"
+      },
+      "targets": [
+        "packages/shared-server/rallar-system/state-sync/validate-state-sync.ts"
+      ]
+    },
+    {
+      "mergeBase": "1e5f5e55e6ff94c016bfe2cc11af92952a30e32f",
+      "source": {
+        "path": "packages/shared-server/rallar-system/services/rallar-rtc-topology-service.ts",
+        "blob": "9aaaff62b0d0e2e05e179a2956dbb68f3db610a8"
+      },
+      "targets": [
+        "packages/shared-server/rallar-system/topology/rallar-rtc-topology-metrics.ts"
+      ]
+    },
+    {
+      "mergeBase": "1e5f5e55e6ff94c016bfe2cc11af92952a30e32f",
+      "source": {
+        "path": "packages/shared-server/rallar-system/topology/replay/create-rtc-topology-work-handler.ts",
+        "blob": "8924cc2d6ff4bf5eb35e1c387e93f3a00f5e8409"
+      },
+      "targets": [
+        "packages/shared-server/rallar-system/topology/replay/finish-rtc-topology-work.ts"
+      ]
+    }
+  ]
+}
